stardadize temmplated items component
